♻️ refactor: separate type in Card component
