+++ packages/card/src/types/context.ts
@@ -1,0 +1,7 @@
+import { CardProps } from './props';
+import { StyleProps } from './styleProps';
+
+export interface ReturnContext {
+  cardProps: CardProps;
+  styleProps: StyleProps;
+}
