+++ packages/card/src/types/props.ts
@@ -1,22 +1,8 @@
 import { AriaAttributes, HTMLAttributes, CSSProperties } from 'react';
+import { FlexToken, Variant, Direction, Size, AsElement } from './base';
 import type { CSSObject } from '@emotion/react';
 import type { ColorToken } from '@jdesignlab/theme';
 
-export interface ReturnContext {
-  cardProps: CardProps;
-  styleProps: StyleProps;
-}
-
-export type CardStyle = 'variant' | 'color' | 'align' | 'justify' | 'direction' | 'size';
-
-export type StyleProps = Required<Pick<CardProps, CardStyle>>;
-
-export type FlexToken = 'start' | 'end' | 'center' | 'between' | 'around' | 'stretch';
-export type FlexMap = Record<FlexToken, string>;
-export type Variant = 'elevated' | 'outlined' | 'filled';
-export type Size = 'sm' | 'md' | 'lg';
-export type Direction = 'horizontal' | 'vertical';
-export type AsElement = React.ElementType;
 export type ExternalStyle = CSSObject & CSSProperties;
 
 export interface CardProps extends AriaAttributes, HTMLAttributes<HTMLElement> {
