+++ packages/card/src/types/styleProps.ts
@@ -1,0 +1,4 @@
+import { CardProps } from './props';
+
+export type CardStyle = 'variant' | 'color' | 'align' | 'justify' | 'direction' | 'size';
+export type StyleProps = Required<Pick<CardProps, CardStyle>>;
